Initial Firefox support in HTML5 offfline app
We are now using an in-memory database loading from and saving to
localStorage as a fallback for browsers not supporting WebSQL like
Firefox.  Performance is expected to be slower on large datasets.
Saving to localStorage is not yet implemented, but loading from
localStorage is. With the present behavior, users can sync with the
server and play with data, but navigating away from the web app,
closing the tab, or reloading will result in loss of unsynced changes.
Allowing for a server sync, no data loss should occur, as upon next
load, data can be fetched from the server.