Initial import of Chrome App
In static/chromeapp, due to directory cleanup, links to javascript in
/static/chromeapp/js may be broken.

client-server schema syncing is partially implemented, so sync is
broken for now.

Nifty javascript-based QR code generator in the chrome app should find
its way to the html5 client

There is some duplication of javascript libraries in the chrome app:
the nature of packaged apps is that they must be self-contained
downloads.  For efficiency, we could link the html5 client to
javascript libraries in the chrome app diractory.  This is messy though
as not all libraries are in the chrome app.  Just a thought.